move cache to core/ctx as it is tighly coupled to the context. Before lived in common, which was misleading
